exiv: Fix type of Pixel{X,Y}Dimension

+++ src/common/exif.cc
@@ -1284,8 +1284,8 @@ int dt_exif_read_blob(uint8_t **buf, const char *path, const int imgid, const in
 
     /* Replace RAW dimension with output dimensions (for example after crop/scale, or orientation for dng
      * mode) */
-    if(out_width > 0) exifData["Exif.Photo.PixelXDimension"] = out_width;
-    if(out_height > 0) exifData["Exif.Photo.PixelYDimension"] = out_height;
+    if(out_width > 0) exifData["Exif.Photo.PixelXDimension"] = (uint32_t)out_width;
+    if(out_height > 0) exifData["Exif.Photo.PixelYDimension"] = (uint32_t)out_height;
 
     int resolution = dt_conf_get_int("metadata/resolution");
     if(resolution > 0)
